Demonstrate slot-scope
And finally... this just shows how you can allow the header/footer
slots in the child (the thing that uses Modal) to have access to
variables from the parent's (Modal itself's) scope.

In this case, it's not actually all that useful -- but the technique
itself does often come in handy.